multim_2636A_ctrlTFT.sh minor cleanups, preparing for better read-based current function abort
git-svn-id: svn://svn.imager.umro/masdax/trunk/scripts/scriptmeas@623 92890ac3-fb6f-4dbc-9a08-888296f68159

+++ multim_2636A_ctrlTFT.sh
@@ -3,9 +3,6 @@
 # $HeadURL$
 # $Id$
 
-#DATACTRLFILE='commtemp/2636_datactrl'
-#SCPIFILE='commtemp/2636_command.scpi'
-
 # source the settings file
 source "$1"
 if [ "$MAINSMU" == "" ]; then
@@ -13,14 +10,8 @@ if [ "$MAINSMU" == "" ]; then
  exit 5
 fi
 
-N=0
-for SMU in $SMUS; do
-N=$(( $N + 1 ))
-SMUdisp[$N]=$( echo "$SMU" | sed -e 's%\.%.display.%' )
-SMUchan[$N]=$SMU
-done
-
-rm "$SCPIFILE"
+[ -e "$SCPIFILE" ] && rm "$SCPIFILE"
+trap "rm '$SCPIFILE' '$DATACTRLFILE'; exit" EXIT
 
 function send_cmd() { # sends command to scpi file
   if [ -e "$SCPIFILE" ]; then
